Refactor utls out into separate files
Also: rename some functions, add extra tests for them, and more comments
