+++ packages/design-tokens/bin/buildCssVarTokens.ts
@@ -3,7 +3,7 @@ import path from "path"
 import { format } from "prettier"
 import * as yargs from "yargs"
 import { defaultTheme } from "../"
-import { makeCSSVariableTheme } from "../src/utils"
+import { makeCSSVariableTheme } from "../src/lib/makeCssVariableTheme"
 
 const { jsonOutput, cssOutput } = yargs
   .option("jsonOutput", {
