+++ packages/design-tokens/src/ThemeManager.ts
@@ -1,5 +1,5 @@
 import { Theme as BaseTheme } from "./types"
-import { makeCSSVariablesOfTheme } from "./utils"
+import { makeCssVariableDefinitionsMap } from "./lib/makeCssVariableDefinitionsMap"
 /**
  * Use this class to set and apply themes, and to access or subscribe to the currently active one.
  * This class fulfills the idea of theming and runtime theme switching by relying on CSS variables,
@@ -50,8 +50,8 @@ export class ThemeManager<Theme extends BaseTheme = BaseTheme> {
     )
   }
   public applyCurrentTheme = () => {
-    const cssVariablesOfTheme = makeCSSVariablesOfTheme(this.theme)
-    Object.entries(cssVariablesOfTheme).forEach(([key, value]) => {
+    const cssVariableDefinitions = makeCssVariableDefinitionsMap(this.theme)
+    Object.entries(cssVariableDefinitions).forEach(([key, value]) => {
       if (this.theme.themeKey === "zen") {
         this.rootElement.style.removeProperty(key)
       } else {
