+++ packages/design-tokens/src/lib/__tests__/themeForTesting.ts
@@ -1,0 +1,17 @@
+export const themeForTesting = {
+  border: {
+    dashed: {
+      borderWidth: "2px",
+      red: "red",
+    },
+  },
+  color: {
+    gray: { 100: "#eee", 200: "#ccc" },
+    purple: { 100: "#eee", 200: "#ccc" },
+    blue: { 100: "#eee", 200: "#ccc" },
+    red: { 100: "#eee", 200: "#ccc" },
+    yellow: { 100: "#eee", 200: "#ccc" },
+    orange: { 100: "#eee", 200: "#ccc" },
+    green: { 100: "#eee", 200: "#ccc" },
+  },
+} as const
