+++ packages/design-tokens/src/lib/patterns.ts
